Avoid Log-Spew When Deciding at Compression-Time not to use an Algorithm
Summary:
Individual compressors are allowed to report themselves as enabled and then,
when asked to actually compress a response, discover that they should not be
used. E.g., when they are enabled by the static config, but disabled by the
request during its execution. Or, in the case of brotli, when it is enabled
for only one of chunked or non-chunked responses, and is given the other type
to compress. In that case, no exceptional condition has occurred, and we
shouldn't log an error.

In the case that constructing a compressor fails, it's the responsibility of
the various `getCompressor()` calls to record that. (Though currently,
they'll simply throw, which won't be caught anywhere here, which is probably
desirable since they're not expected to ever fail except as a result of
OOMing.)

Differential Revision: D12927652

fbshipit-source-id: 6b494b92e9699b0560a1ed251a59e1b78f2b281f