Always use ThrowOnCallByRefAnnotationMismatch in Rx body.
Summary:
The fallback calling convention is not compatible with Rx bodies: it introduces reffy bytecodes at every function call site.
Fix this by keeping track of whether we're emitting an Rx body or not, and if we are then unconditionally use the new calling convention (regardless of if ThrowOnCallByRefAnnotationMismatch is enabled). Rx code is not allowed to pass by reference, so it should already always satisfy this stricter requirement.

Reviewed By: jano

Differential Revision: D13254772

fbshipit-source-id: e88ef8a210d04cbb463f55fccfa18e9afebf3a34