Fix assert caused by emitCheckSurpriseFlagsEnter()
Summary:
emitCheckSurpriseFlagsEnter() emits code into "v", sets "v" to the done
block, sets "vcold" to the surprise block, and then emits a vinvoke into "vcold"
(closing the block). However, "v" and "vcold" might be the same Vunit, meaning
that both "v" and "vcold" might still be pointing at the surprise block. The
callers do not expect this, and when they try to insert more code into "v",
we'll assert because the block is closed. Fix this simply by setting "v" to the
done block after emitting the vinvoke, not before.

Reviewed By: markw65

Differential Revision: D13290325

fbshipit-source-id: 455f004c55dba1c692fed819b8cbbd678dc2adf0