Save SQLite dep table: Don't look for existing rows unnecessarily
Summary: Based on the "before" log, it takes more than 10 seconds to look up existing rows when we're saving from a full init. Makes saving a tiny bit faster (3-4 seconds on 200K files) - doesn't translate into the savings of 10 seconds, but it's still a little faster and also makes more sense as you read the code.

Reviewed By: dabek

Differential Revision: D13207095

fbshipit-source-id: 561061067159eae7e36dfbe4efba9301307729de