Propagate reified generics through Aast for function calls
Summary:
Created a new type in Aast for generics, which just pairs a hint and a bool, and changed the node that represents function calls to contain a `generic list` instead of a `hint list`. Also renamed `hl` to `gl` in the typechecker and used List.unzips to pull out the hint list to make explicit in the dispatch_call section. Finally, forced localization of fun_types to keep arguments when a reified generic appears in the parent, because all type arguments (including erased) must be spelled out in that case.

The rest was just fixing build errors.

Reviewed By: oulgen

Differential Revision: D13245185

fbshipit-source-id: 55c0ff46e12c78d57aa85172edf445c6bfc3506c