Adding new decl for redeclared method
Summary:
Adds a redeclared method to the decl map as if it were a normal method declaration in the class. In the test case, note that `C::f` ends up with `T2::f`'s type, while `C::g` gets `T::f`'s type.

The effect is similar to having two traits with a name conflict and a class method to override them, but stronger in the sense that method redeclaration removes the inherited method too.

Typing is still unaware of this removal, so another diff is required to clean up the error message in the test case.

Reviewed By: jamesjwu

Differential Revision: D13071433

fbshipit-source-id: eeac21d05799253a07525f63ea9c1b39a52c4fb4