Simplify subtype: unions on right
Summary:
The main remaining rule that we don't implement in `simplify_subtype` is unions on the right. Let's do it now. It's sound to reduce
```
t <: (t1 | t2)
```
to the disjunction
```
(t <: t1) \/ (t <: t2)
```
but not complete e.g. consider
```
(t1 | t3) <: (t1 | t2) | (t2 | t3)
```
But as we deal with disjunctions on the left first (reducing the example above to `(t1 <: t1 | t2) /\ (t3 <: t2 | t3)` we know that the precise situation above won't arise.

TODO: determine whether we've have lost completeness in some other circumstances.

Reviewed By: CatherineGasnier

Differential Revision: D13276093

fbshipit-source-id: 60cd87e5425432cf33e8da5a640662c4e7bf9fbe