Create a subtyping constraint between the expanded types in the default case of simplify_subtype under the new inference flag
Summary: What it says on the tin.  We cannot do this in the legacy inference algorithm because an unresolved type and a type variable bound to an unresolved type are handled differently.  This looks like an optional optimization in the new inference algorithm, but it actually appears to be necessary; see the test case below.

Reviewed By: andrewjkennedy

Differential Revision: D13179202

fbshipit-source-id: 5fec7467f12c72250dbdb8f4d5a3d0b5a31f1627