Don't add again an already existing upper/lower bound of a type variable
Summary:
Suppose that we have an environment with the following `tvenv`:
```
```
(such an environment arises, e.g., in the new test case added below before the `return` statement), and suppose we apply `prop_to_env` to the subtyping proposition `#1 <: int` in this environment.  Then we enter an infinite loop: we want to add `int` as an upper bound for `#1`, but before that we impose an additional constraint, which should hold by transitivity of subtyping, that `#2 <: int`, which implies, again by transitivity, that `#1 <: int`, and we are back to the proposition from which we started.  As the result, the test case below overflows stack and causes a segmentation fault.  A simple (but probably not perfect) fix is to check if a type we are about to add as an upper or lower bound of a type variable is already contained in the set of upper resp. lower bounds.

Reviewed By: CatherineGasnier

Differential Revision: D13196287

fbshipit-source-id: d3ade5e40d3ea0d0d084012a50c3cb89b01231be