Collapse equivalent type variables in tvenv
Summary: Instead of avoiding creating cycles between equivalent type variables when moving propositions `v <: t` and `t <: v` to `tvenv`, I'm suggesting to try and collapse cycles immediately after creating them.  The reason is that if we want to avoid creating cycles we still need to apply transitivity to the constraints arising from the bounds on the variables in the cycle.

Reviewed By: andrewjkennedy

Differential Revision: D13234272

fbshipit-source-id: b97b777af075c1d552d21c4716df305c57c6c2ec