Properly detect generic parameters in statics under constraint solver
Summary: Under the `--new-inference` flag we have disabled the use of the "todo" list from the environment. This makes sense for subtype assertions generated off `as`, `super` and `where` constraints, but there is another use that is not so easy to express in constraints: detecting the use of generic type parameters in the types of statics and globals. For that, I've re-enabled the use of the to-do list. It's still a hack, and a better long-term solution would be to accumulate the statics/globals declared by a function, or simply ban them altogether.

Reviewed By: Wilfred

Differential Revision: D13319988

fbshipit-source-id: 061d04bf31fdc835957f51e610d702a64995d967