Fix gen-type-scanners view of static defs for clang debug info
Summary:
Static definitions for member variables are defined in a
compilation unit that may have an incomplete view of the type. This
incomplete type will have a type signature to refer to the complete
definition in a separate type unit. debug-parser-dwarf.cpp already
supports this, but it has the restriction that the type declaration in
the cu must appear before its static member variable definition,
otherwise local_mappings will not properly reroute this static
definition back to the original type struct definition when the
definitions is first seen. HHVM built with clang emits the static
definition before the type declaration. This diff changes
debug-parser-dwarf to remap all static defs after genNames is complete
and all local mappings have been established. It also changes the
static definitions from unordered to ordered map because the latter
performed better in my benchmarks.

Reviewed By: markw65

Differential Revision: D13240332

fbshipit-source-id: f66200129b57ad8b75f237c5b0674c136d76263d