Implement VregSet
Summary:
Implement VregSet, a set of Vregs. Unlike Regs, there's potentially an
unbounded number of Vregs, so we cannot use a fixed size bitset. A variable
sized bitset can become expensive if the universe of Vregs is large and the set
is sparse. Instead, a sorted vector of Vregs gives the best overall
performance. Also add some exhaustive test cases for it.

Reviewed By: mofarrell

Differential Revision: D13230811

fbshipit-source-id: 92e1e384d4de8bce7d6fa55f1d549fea36008ebc