Implement sub_string using subtyping
Summary:
Hack's concatenation and string interpolation operations require that their arguments satisfy the `sub_string` predicate defined in `Typing_subtype`. Unfortunately this predicate is not based on subtyping and so doesn't leverage any of the infrastructure built for that purpose. With constraint-based type inference in particular, we are hitting the `Tvar` case when a type variable has yet to be "solved".

It turns out that it can be implemented in terms of sutyping if we use a union type: we want to test that the parameter type `t` is a subtype of `?(arraykey | bool | float | resource | dynamic)`, which includes all of the cases currently considered by `sub_string`. (With support for the legacy `Stringish` behaviour we add the fake `Stringish` class to the union.)

This diff uncovered a glitch in the constraint-based support for unions on the right of the subtype relation: if `num` is on the left, so we're checking something like `num <: t1 | ... | tn` then we should first expand it to a union of "atomic types" i.e. `int|float`.

Reviewed By: kmeht

Differential Revision: D13321932

fbshipit-source-id: 7961dc7c61e52bebfd338442d92cf724ebed05cc