Fix type-checking of suspend with constraint-based inference
Summary:
When type-checking function and method calls, we keep track of whether the call appears under `suspend` or not, and correspondingly we assert whether the function must be a coroutine or not.  The check whether the function is a coroutine is performed by case analysis on the function type.  Furthermore, the existing implementation assumes that the function type is expanded and cannot be a type variable.  These assumptions no longer hold in the constraint-based inference algorithm.

A morally right way to resolve this issue would be to replace the case analysis with a subtyping check: we would introduce a function type `Tfun ft_super` of which the given function type `fty` must be a subtype.  The function supertype would represent a type `(function(v1, ..., vn): v)`, where `v1`, ..., `vn`, `v` are fresh variables, and the number of variables `vi` would be determined based on the total number of parameters passed to the function whose call we are type-checking.  `ft_super` would have the field `ft_is_coroutine` set based on the knowledge of whether the call appears under `suspend` or not.  The subtyping check would enforce that `fty` is a function type (possibly unresolved) whose `ft_is_coroutine` matches `ft_super.ft_is_coroutine` (for unresolved function types that would mean that all functions comprising the unresolved type have the same value of `ft_is_coroutine`).  This check would also deal correctly with type variables.

(Aside: we could use "empty" type `Tunresolved []` for `vi`s and `mixed` for `v` because we only care about whether the value of `ft_is_coroutine` is correct.)

However, the above approach is hard to implement at the moment because function types have many other attributes, e.g., dealing with reactivity / mutability and disposable types.  The latter present a particular problem because two function types that are in a subtyping relationship must share the same value of `ft_return_disposable`, which means we cannot come up with the right `ft_super` without inspecting `fty`, which defeats the whole purpose.  I don't fully understand this business of disposable type and it's not a good time to change it.

Therefore, I'm going with an inferior (and hopefully temporary) solution: we expand the function type and if we still get back a variable, we check if all its lower bounds are coroutines.  The morally right approach would add the manufactured function type, `Tfun ft_super`, as an upper bound to the type variable standing for the function type, which in particular would assert by transitivity that every lower bound of that type variable is a subtype of `Tfun ft_super`.  This upper bound would stick around and we would check against it if the function type "grows" (acquires new lower bounds).  This doesn't happen with the current implementation and I suspect it might suffer from the problem similar to "TODOs are checked prematurely": if the type variable has no lower bounds, we will consider it non-coroutine, but a constraint at a later point in the program may add a coroutine as a lower bound to the type variable.

Reviewed By: andrewjkennedy

Differential Revision: D13339750

fbshipit-source-id: ac9c0749de0134ab45795985a8327c6998c9b11e