Prevent IDE_IDLE spam
Summary:
D13088232 introduced "tick_showStatus" loop which once a second (in Main and Lost_server) will send a showStatus request to the client. The client naturally responds with a response once a second.

The way hh_server works is if it ever receives a request over the persistent connection, then it won't accept any CLI requests until the persistent connection has sent IDE_IDLE. Therefore, it's crucial that IDE_IDLE gets sent enough. (but it doesn't harm correctness to send additional IDE_IDLE messages).

This is a tough guarantee to fulfil!

The way clientLsp fulfils it is, if ever it receives any message from the server, or any message from the client, then after handling it it always indicates that it must send an IDE_IDLE on the next available idle tick. This means we don't need to keep careful tally of whether we've sent any requests to the server or not that would need IDE_IDLE.

BUT

Because of tick_ShowStatus sending a request to the client once a second, that means we got a response back from the client once a second, which (with the logic above) required an IDE_IDLE to be sent to the server on the next available tick.

End result: the server got zillions of IDE_IDLE messages, and they kind of blocked up all our logging.

This diff changes it so that request_showStatus will only actually send a request to the client if the content of this message is different from the previous one that was sent to the client.

Reviewed By: arxanas

Differential Revision: D13355957

fbshipit-source-id: 03deba93167b1fce29fc6cb7d5982d57ecd30fd4