Fix segfault in backtrace hash computation for returning functions
Summary:
When a function is returning and its locals are already decrefd, we may invoke
surprise check, which may call arbitrary PHP code, which may try to compute
backtrace hash, which accesses metadata on the frame stored in already freed
local.

Check for this condition, like we already do in all other backtrace APIs.

Reviewed By: alexeyt, markw65

Differential Revision: D13335015

fbshipit-source-id: a6bd61c75bbefa18cd68a10457537b72275fbc81