Typing rules for Shapes::removeKey
Summary:
What it says on the tin.  One caveat: we don't yet have formalization of inout function parameters in the spec, so instead of treating `Shapes::removeKey` as a side-effecting operation that mutates the passed in shape, I'm viewing it as a pure function that returns a new shape with the supplied field name removed, so that
```
Shapes::removeKey(inout $s, $field_name);
```
in Hack is equivalent to
```
$s := Shapes::removeKey($s, $field_name);
```
in the spec.

Reviewed By: andrewjkennedy

Differential Revision: D13304240

fbshipit-source-id: 877252a59e45aee46b2a27fcb0907554d4cab526