Keep track of unset fields in an open shape type
Summary:
The current representation of open shapes in the spec doesn't quite faithfully formalize open shapes as implemented in Hack.  We need to keep track of fields unset using `Shapes::removeKey`.

This diff changes the representation of open shapes and adjusts the existing rules (and, in particular, adds a new subtyping rule).  The typing rules for `Shapes::removeKey` are given in stacked diff D13304240.

Reviewed By: andrewjkennedy

Differential Revision: D13304235

fbshipit-source-id: 9e9bb328b0c4c5e1014e98c8f6ba2837b3fb8d62