Assume hhi files are in decl mode
Summary: The previous diff removed the `disable_decl` option, meaning any files marked as decl would be interpreted as partial mode. This diff removes the need to specify decl in `.hhi` files, and instead assumes this mode via the file extension. The next diff will remove the ability for users to specify decl altogether, and the mode will be repurposed for exclusive use in `.hhi` files.

Reviewed By: Wilfred

Differential Revision: D13347442

fbshipit-source-id: 9d9b5231bcb86a770493ba86b03046ca0fbeaef0