Fix the typing of Shapes::removeKey in the constraint-based type inference
Summary:
When type-checking a call to `Shapes::removeKey`, say
```
Shapes::removeKey(inout $s, 'foo');
```
we "shrink" (by removing the field name `'foo'`) the type of`$s` and set the type of `$s` in the local environment to the shrunk type.  The shrinking works by case analysis on the type: if the type is a shape, then we remove `'foo'` from the set of field names, and if the shape is open we also record that the field name has been unset; if the type is unresolved, we shrink each of the constituent types; otherwise we leave the type as is.  This doesn't handle the case of a type variable (the legacy algorithm expands the type and assumes that the case of a type variable will never occur).

At first, I thought we could represent this shrinking as intersection with `shape(...)(unset fields: foo)`, but that's not correct.  In particular, when the shape is closed, there is no subtyping relationship between the shape and the shape with the field name removed.  Therefore, I resorted to an uglier but more direct approach: for a type variable, we shrink all its bounds (both lower and upper).

There is an interesting wrinkle: to avoid aliasing problems, we need to create a fresh type variable corresponding to the shrunk type.  I've added a test illustrating this.  See inline comments below.

Reviewed By: andrewjkennedy

Differential Revision: D13320519

fbshipit-source-id: 788bfa6cae9e41263eff843c47e6f11e0384e4ee