Fix segfault when reflecting over trait properties
Summary:
Reflection was trying to load trait properties from PreClass of a Class that
used the trait, but in non-repo mode they exist only on a PreClass of a Class
representing the trait.

Fix this by keeping track of a pointer to PreClass::Prop from each Class::Prop
and Class::SProp. Eliminate memory impact by removing docComment and userType,
which is cold and can now be accessed via preProp.

There's probably opportunity to remove more stuff from Class::{,S}Prop, such as
TypeConstraint and RepoAuthType, but that's outside of scope of this diff.

Reviewed By: ottoni, alexeyt

Differential Revision: D13333958

fbshipit-source-id: 74829a233dcbf7ff5bb252362014a116d40f80e9