Instantiate type parameters during linearization
Summary:
Before this diff, the added test case produced this linearization:

```
Linearization for class \A:
[\A(child)]
Linearization for class \B:
[\B(child), \A<bool,T1>(parent)]
Linearization for class \C:
[\C(child), \A<string,T1>(parent)]
Linearization for class \D:
[\D(child), \B<float>(parent), \A<bool,T1>(parent), \C<int>(parent), \A<string,T1>(parent)]
```

Note that `\A<bool,T1>` and `\A<string,T1>` both occur in the linearization. Without knowing in what class the type parameter `T1` was declared, we do not know what type to substitute for it when instantiating `\A`.

This diff resolves this issue by instantiating type parameters during linearization. After this change, the added test case contains `\A<bool,float>` and `\A<string,int>` in its linearization (eliminating the ambiguous generic types named `T1`).

Reviewed By: arxanas

Differential Revision: D13282458

fbshipit-source-id: be641e0d2fe03c7d9bf654440ba3fbbca658f534