Fix FPushCtorS to work with reified generics for new self and parent and throw on new static
Summary: On D9414305, we left `FPushCtorS` unmodified while passing the reified generics for class instantiation. This diff adds support to `FPushCtorS` to allow using `new self` and `new parent` inside a reified class and have the new instantiation also be reified with whatever the original class was reified with. We do this solely through `FPushCtorS` instruction by disabling reduction to `FPushCtorD` when the class name is known statically. This diff makes it so that `new static` on a reified class throws a runtime error.

Reviewed By: paulbiss

Differential Revision: D10366328

fbshipit-source-id: cbbf0ebdce8b14f0e2951eab0d3390c9c96e965d