Add memo eager async optimization
Summary:
Add support for optimizing async memoized functions better. Until now, a
memoized async function couldn't take advantage of the eager return
optimization. IE, all of the memoized values would be wrapped in wait-handles
and be returned as such. This means callers of the memoized function would
always received wait-handles and have to unwrap them.

Instead, have the memoize wrapper call the async wrapped function opting into
the eager return optimization. If the wrapped function returns eagerly, store it
in the memo cache like normal. If it returns a wrapped wait-handle, store that
in the memo cache as well, but also store a bit which indicates it is a
suspended wait-handle. We use a new bytecode called MemoSetEager for this
purpose (indicating an eager value), or MemoSet (for the suspended case).  We
use the aux bits to store whether the value is suspended or not.

When the value is retrieved, we use a new bytecode called MemoGetEager. This is
similar to MemoGet, but can branch to two different locations. The second branch
is if the aux bits indicates the value is a suspended wait handle. If the
instruction fallsthrough, its an eager value. Finally we use a new RetCSuspended
bytecode to return the value in the suspended case. This behaves like RetC,
except it does not try to wrap the returned value in a wait-handle (because the
returned value is already wrapped in one).

Reviewed By: jano

Differential Revision: D13072566

fbshipit-source-id: 234f72354630e4eb3fa164218d4d892e39c11bfb