Simplify unions and nullables subtyping a type variable
Summary:
Currently we don't reduce `t <: v` in `simplify_subtype` except when `t = v`. So `t` ends up as a lower-bound on `v` in `tvenv`, and if `v` is covariant, we attempt to unify `v` and `t`. This is problematic if `v` occurs in `t`.

In the special case of unions, and nullables (which are really unions), we *can* simplify further: we know that `t1 | t2 <: v` iff `t1 <: v` and `t2 <: v`. So that's what we do here, also for the case of nullables.

Reviewed By: manzyuk

Differential Revision: D13397379

fbshipit-source-id: 3e097ab400918ebc64dfdb79e82ee81e7e6d4abe