Add -fdebug-types-section to hhvm gcc builds
Summary:
The .debug_info section in our non-lto builds is getting very close to
4G, at which point the build fails.

Applying -fdebug-types-section to the hphp directory drops the size to
about 3G which gives us plenty of headroom. Unfortunately, there are a
number of known bugs (either compile time crashes, or link time build
failures) when applying that flag, so although it currently works, I
don't want to commit this unless we have to (and hopefully after the
bugs are fixed).

Putting this up in case there's no other solution...

Also, we can't apply this in lto builds, because compilation
insta-crashes.

Reviewed By: ricklavoie

Differential Revision: D13343690

fbshipit-source-id: 7ff1e670eafd55c53ecf737c03d6ad60c877fdbf