Type inference for array get operation when array has Tvar type
Summary:
The array access syntax (square brackets) is highly overloaded. This makes it hard to type check the operation `e1[e2]` when `e1` has a type variable type. But it is possible to do so in cases where the operation can be simulated by a generic function
```
function array_get<Tk, Tv>(KeyedContainer<Tk, Tv> $array, Tk $key): Tv;
```
This is essentially the approach we take here. This does *not* deal with shape or tuple access.

Reviewed By: CatherineGasnier

Differential Revision: D13397953

fbshipit-source-id: ef342047fdbbf5c49cddc4ad5337e16feaf3c144