Create module to map from AST to NAST
Summary:
This is a small helper module to get from the legacy AST (AST) to an annotated AST (AAST), which in this case will be the named AST (NAST).

This is so we can quickly get a NAST for codegen, without going through the naming phase. It also makes it easier to adapt the AAST structure and see where the differences lie between the AST and AAST.

This module is designed to be temporary and will be deleted once we do compilation and type-checking solely on the AAST.

Reviewed By: Wilfred

Differential Revision: D13320218

fbshipit-source-id: 6727c038a7bb3cb39af99ed57366030397c4d139