Gate eager solving of invariant type variables behind a flag
Summary: In the diff above this one, we introduce "eager" solving of invariant type variable constraints for constructs such as method invocation. This is something of a departure from the "don't close off solutions" approach to constraint solving so it would be nice to be able to opt in and out in order to experiment with it. Hence the flag.

Reviewed By: kmeht

Differential Revision: D13417345

fbshipit-source-id: 02a602f8e3dbb6c4ca2a14bb0d20bd9830e9e465