Pre-emptively solve if accessing member on a type variable
Summary:
Currently we solve type variables only if we are sure that the solution is the most general i.e. it won't prevent further valid typings. This is *not* the case if it appears invariantly (i.e. both covariantly and contravariantly) in the expression type in which it was introduced.

But this means that we sometimes reject code because the type-checker needs to do case analysis on a type, and it just sees a type variable. The most common case is property access or method invocation. In these situations, because we're going to reject the code anyway (because there isn't enough type information to resolve the access), we may as well try and solve for the type variable, even if it's invariant. Highly-overloaded array operations are another example.

We gate this eager solving on the `--new-inference-eager-solve` option, in order to experiment with and without it.

Reviewed By: manzyuk

Differential Revision: D13415244

fbshipit-source-id: 0b2195e723aa4a41c65e39a665e288ed4fee243d