Initial commit of graph-coloring register allocator
Summary:
Introduce partial skeleton of the new graph-coloring register
allocator. Its usage is controlled by a new option "UseGraphColor" which
defaults to false for now. Even when enabled, its limited to optimized
translations because of speed reasons.

Due to the complexity of the allocator, it is being introduced in pieces. This
diff adds the basic skeleton of the allocator, each diff will add a new piece of
functionality. The allocator won't work until the last piece, so assert if you
try to run it for now.

Reviewed By: mofarrell

Differential Revision: D13300490

fbshipit-source-id: fe159d8874b3897453aa74066312f0d6288e15ca