Add flag to disable intish cast
Summary:
What it says on the tin. We want a flag to disable intish cast.

I wanted to use the IntishCast enum to mark the JIT helper templates, but it has slightly different meaning. IntishCast::AllowCastAndWarn means that the decision should be up to the runtime flag, since the goal of these template parameters is to allow callees to force the intish cast to occur in the first place.

On the other hand, the purpose of the termplate parameters on the JIT helpers is to lift the check for a runtime flag up into the code generation. I added a ICMode enum to represent the value of this parameter, so we don't get these two mixed anywhere.

Reviewed By: mxw

Differential Revision: D13110799

fbshipit-source-id: 560741c44fa72433a339edead39e22133e3e562e