hh_server should output the number of edges added in JSON mode
Summary:
I need this for testing in a subsequent diff. `hh_client` already outputs the # of edges added when saving state, but `hh_server` doesn't.

NOTE: there is an inconsistency in how hh_client handles this.

The discrepancy now is that `hh_client` doesn't output the errors, if there are any, and --gen-saved-ignore-type-errors is specified:

```
> hh_client --save-state ~/foo/bar --json --gen-saved-ignore-type-errors

{"result":3623}
```

Otherwise, it outputs this:

```
> hh_client --save-state ./foo/barft --json

{"error_message":"There are typecheck errors; cannot generate saved state."}
```

Reviewed By: arxanas

Differential Revision: D13056090

fbshipit-source-id: b1487f5a050d7d0b3bcd573a22aee1d72ee5f39c