Add optional verification of Rx function bodies.
Summary:
Add the RxVerifyBody option, use it to enable warnings or hard failures in the verifier when bytecode-level invariants in Rx function bodies are violated.
Add unit tests that cover all the bytecodes that we want to reject (including member operations, which we reject conditionally based on operands and the state of the minstr sequence as a whole).

Reviewed By: paulbiss

Differential Revision: D13254773

fbshipit-source-id: 8a194ed542e8e1ceb19ba370c072cf97e673e6fb