Add pseudo conversion support to graph-color reg-alloc
Summary:
The graph-color register allocator needs to put the entire unit into
SSA form before it can function. This means rewriting Vregs in the instruction
arbitrarily. However, some instructions have RegSets as operands, or they have
implicit register effects (which aren't explicit in the instruction. These
cannot be rewritten easily because RegSets can only contain physical registers
(and the implicit effects aren't stored anywhere).

Deal with this by inventing a number of "pseudo" instructions. These
instructions have no inherent meaning, but are meant to serve as stand-ins for
the problematic instructions. There's a variety of them, covering the different
categories of instructions that the register allocator needs to distinguish
between. All the physical register operands are converted to equivalent Vregs
and stored as operands in the pseudo instructions. After this, the pseudo
instructions can be manipulated (rewriting Vregs) like any other
instruction. This lets us put everything into SSA form.

Once register allocation is done, the pseudo instructions are converted back to
their original instructions. The "pos" field in the instruction is used as a key
to know which instruction to convert them back into.

Reviewed By: mofarrell

Differential Revision: D13308502

fbshipit-source-id: 5cf74810840445b683400a09c32d33c9466416c4