Implement __SoftLateInit
Summary:
Implement a new property attribute called __SoftLateInit. This attribute is
similar to __LateInit, except instead of throwing, it raises a notice and sets
the property to a configerable default value. Since the property is set, another
access will silently just return that value, as if it had been set manually to
that.

Add a builtin called HH\set_late_init_default_value() to configure what
__SoftLateInit should use.

Reviewed By: markw65

Differential Revision: D13472371

fbshipit-source-id: 1bab7a196b98f025872c32ce5e9578bf7e7c6ba7