Fix assertion triggered in headers_list
Summary:
The various ArrayInit builders cannot be given an underestimate since they are
designed to (and enforce in dbgo builds) never escalate the array its
building. headers_list() was passing a potential underestimate, so we might
assert if it proves wrong. Since this isn't a hot function at all, just use
Array instead of VArrayInit.

Reviewed By: oulgen

Differential Revision: D13484960

fbshipit-source-id: a5ef490e256a3396f5edd8a08c194b988424530e