Improve error messages for references
Summary:
The rules around reference usage are subtle in Hack. Improve the error messages to help users (and us!) understand when they can use `&`. For errors that aren't controlled by flags, remove "no longer", and keep wording concise and consistent.

References are generally banned. The only exceptions are:

* Function calls (although references are still banned for special functions like `isset`)
* Code in partial mode that isn't using arrays, `<<__Rx>>` or `--disallow-assign-by-ref`

Add comments to clarify the logic in the code.

Reviewed By: periodic1236

Differential Revision: D13380023

fbshipit-source-id: 5a67475b17e44f3c1464f80a09372b8a088460f7