Make closures inherit the Rx level of their scope in the emitter.
Summary:
In the emitter, let closures of all types inherit the rectivity of their scope unless overridden.
Do this by distinguishing explicit __NonRx from no annotation, tracking the declared reactivity of scopes in Ast_scope, capturing the Rx level of the enclosing scope of every closure (where it's not NonRx) in the Emit_env's global state, and defaulting the Rx level of closures that don't have an explicit annotation to that of their scope.
Add a unit test for this behavior.

Reviewed By: vladima

Differential Revision: D13438662

fbshipit-source-id: bf786630bffe81d921bdaf28b3b7e8f6b4445b37