Implement constant materialization in vasm-graph-color
Summary:
Implement the materialization of constants in vasm-graph-color. That is, turning
uses of Vregs representing constants into normal Vregs defined by a ldimm*
instruction. While this can be done very trivially by inserting the constants at
the beginning of the unit, instead use dataflow to compute a more optimal
placement. This finishes the "preparation" stage of the algorithm.

Reviewed By: mxw

Differential Revision: D13311853

fbshipit-source-id: e32157d65636b9f633d809f634f6257ee91a3e71