Add register class inference to vasm-graph-color
Summary:
The register allocator needs to know which set of physical registers a Vreg can
be allocated to. Analyze a Vreg's uses and defs to constrain it a particular
register class which determines which physical registers it is eligible for. If
the uses and defs are incompatible, insert copies as necessary to resolve the
incompatibility.

Reviewed By: mxw

Differential Revision: D13318038

fbshipit-source-id: 1de482d38bd50dca307a78059b8bddb96d8c1977