Add class variable finality and visibility
Summary:
Add finality and visibility to class_var.

I'm a bit confused about when final could appear on a class var since as far as I know, this is a parse error and also not allowed in php? I can also put up a diff to wipe this completely if people think that is cleaner.

Reviewed By: arun11111

Differential Revision: D13470460

fbshipit-source-id: 30ab9b10fce419ab13087dd47599fd1abd8d92ce